Moved methods in MathHelper that are used by the API to Util.
