+++ src/main/java/obsidianAnimations/CommonProxy.java
@@ -7,7 +7,6 @@ import net.minecraft.entity.EntityList;
 import net.minecraft.util.ResourceLocation;
 import obsidianAPI.registry.AnimationRegistry;
 import obsidianAnimations.entity.EntityDummyPlayer;
-import obsidianAnimator.ObsidianAnimator;
 
 public class CommonProxy
 {
